+++ test/public/order-test.rb
@@ -2,31 +2,10 @@ require 'minitest/autorun'
 require 'sisimai/order'
 
 class OrderTest < Minitest::Test
-  Methods = { class:  %w[make default another] }
+  Methods = { class:  %w[make another] }
 
   def test_methods
     Methods[:class].each { |e| assert_respond_to Sisimai::Order, e }
-  end
-
-  def test_default
-    cv = Sisimai::Order.default
-    assert_instance_of Array, cv
-    refute_empty cv
-
-    cv.each do |e|
-      assert_instance_of String, e
-      assert_match /\ASisimai::Lhost::/, e
-
-      cf = './lib/' << e.gsub('::', '/').downcase + '.rb'
-      assert_equal true, File.exist?(cf)
-      # assert_equal true, require(cf)
-      # assert Module.const_get(e)
-    end
-
-    ce = assert_raises ArgumentError do
-      Sisimai::Order.default(nil)
-      Sisimai::Order.default(nil, nil)
-    end
   end
 
   def test_another